change schema from bigint to number for k and v

+++ wr-register/src/module_bindings/delete_register_reducer.ts
@@ -11,5 +11,5 @@ import {
 } from "spacetimedb";
 
 export default {
-  k: __t.i64(),
+  k: __t.i32(),
 };
